Added a 'join' and 'take' primitive for joining 2 streams into a stream of tuples. There's something a bit unsatisfying with this, as I can't express which element 'hasChanged'. Maybe I'd need my own tuple classes. Ugh.
got generalised 'reduce while true' stream working.
This allows us to emit a sequence of reductions, with each reduction emitting when a boolean stream transitions from true->false.
This is a building-block for being able to do time-segment reductions (e.g. accumulate accVol during market hours and then emit on the close).

Shame I couldn't simply use 'join' and 'filter' for this (due to being unable to tell which element in the tuple has fired). Ok, I could have constructed a tuple stream of fire states, but that is just getting really messy for unecessary purity.