Still trying to harmonise a system of requirement that yield expected and consistent semantics for bucketing / reducing. I've had to add some graph features: 1. instant edge removal, this is so that I can stop one cell receiving a value, and designate another cell as the target before a source fires its event. To do this safely, I have to know that the currentFiringNode has a causal relationship with the listener target. 2. a fix to hasChanged, to reflect correct hasChanged semantics in the face of cycles.
Fixes to tests, I've got SliceTests and its 'old style' variant working again, but some of the single-terms tests are now broken.
I have more confidence in the correctness of the SliceTests, so will work on the others now.

I'm also having second thoughts about the concept of firing calculate on a bucket whenever a mutator is called by the binding machinery, it makes for some confusing semantics and safety detection when the bucket is able to wire itself to other functions, and generate its own events (I need to support this, imagine a Bucket that has autowired itself to a 1 minute timer and doesn't fire for other events). If someone wants a fire after mutation, they can always call wakeup this cycle.
Of course, Buckets should still be functionally 'after' the source of mutations